+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### [14.1.0](https://kaos.sh/ek/14.1.0)
 
-- **`[req]`** `PostFile` replaced by `SendFile`
+- **`[req]`** `PostFile` replaced by `SendFile` with support of different HTTP methods
 - **`[req]`** Added method `Header.Has`
 
 ### [14.0.0](https://kaos.sh/ek/14.0.0)